Moved the modal of UsersWhoLiked into a separate shared component and it used in Post and Comment
