+++ src/frontend/src/components/Main/Feed/Post/PostButton/NewCommentInput/Comment/UserWhoLikedComment/UsersWhoLikedComment.js
@@ -1,33 +1,12 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import Typography from '@material-ui/core/Typography'
-import styles from './styles'
-import UsersWhoLikedCommentItem from './UsersWhoLikedCommentItem'
-import ThumbUpAltIcon from '@material-ui/icons/ThumbUpAlt'
+import UsersWhoLiked from '../../../../UserWhoLiked/UserWhoLiked'
 
 const UsersWhoLikedComment = (props) => {
   const {usersWhoLikedComment, commentId} = props
-  const usersWhoLikedThisComment = usersWhoLikedComment[commentId]
-  const classes = styles()
 
   return (
-    <div>
-      <div className={classes.title}>
-        <Typography variant="h3" className={classes.subtitle}>
-          All reactions
-        </Typography>
-        <div>
-          <ThumbUpAltIcon className={classes.likeMini}/>
-        </div>
-        <Typography variant="h5" className={classes.number}>
-          {usersWhoLikedThisComment.length}
-        </Typography>
-      </div>
-      <hr className={classes.line}/>
-      <div className={classes.usersWhoLiked}>
-        {usersWhoLikedThisComment.map(user => <UsersWhoLikedCommentItem key={user.id} user={user}/>)}
-      </div>
-    </div>
+    <UsersWhoLiked id={commentId} usersWhoLiked = {usersWhoLikedComment} />
   )
 }
