+++ src/frontend/src/redux/Modal/modalReducer.js
@@ -13,7 +13,7 @@ import {
   EDIT_EXPERIENCE
 } from './modalTypes'
 import AddNewPost from '../../components/Main/Feed/AddNewPost/AddNewPost'
-import UsersWhoLiked from '../../components/Main/Feed/UserWhoLiked/UsersWhoLiked'
+import UsersWhoLikedPost from '../../components/Main/Feed/Post/UserWhoLikedPost/UsersWhoLikedPost'
 import AddEducationModal from '../../components/ProfilePage/ProfileModal/AddEducationModal/AddEducationModal'
 import AddNewCertificationModal from '../../components/ProfilePage/ProfileModal/AddNewCertificationModal/AddNewCertificationModal'
 import EditEducationModal from '../../components/ProfilePage/ProfileModal/EditEducationModal/EditEducationModal'
@@ -61,7 +61,7 @@ const modalReducer = (state = initialState, action) => {
           break
 
         case USERS_WHO_LIKED_POST:
-          content = <UsersWhoLiked/>
+          content = <UsersWhoLikedPost/>
           break
 
         case USERS_WHO_LIKED_COMMENT:
